Mothballed argument processors. Added new bias command/

+++ src/bubo/commanding/commands/rotorcommands/SetRotorBiasCommand.hpp
@@ -25,7 +25,7 @@ public:
 	virtual void execute() const;
 
 private:
-	long bias;
+	uint16_t bias;
 };
 
 } /* namespace commands */
